requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
argparse==1.2.1
python-ldap==2.4.18
wsgiref==0.1.2
